Format code with fourmolu
This unifies different coding styles that were present in the
repository. The drawback is that when developing you usually cannot
apply fourmolu (or any other formatter, for that matter) to .hsc files
properly, as it cannot reasonably handle CPP.

But manually formatting regions that have changed seems like a good way
to go.