Fix build on macOS
More precisely, the "hole" seems to be the combination of configuration
parameters where we (are not on Windows and) we lack wide character
support one way or another.

I'm pretty sure that this is nothing that actually happens anymore
nowadays, but might be due to a misconfiguration or missing libraries on
the macOS CI machines.

Dropping the CWString and only using built-in functions in
Foreign.C.String (thus letting 'base' do the platform-specific work for
us) might help as well, if feasible.